Adding a new unit test and fixing up regexes
The regexes for the name fields (composer, lyricist, etc.) have been unified for simpler maintenance.  The regex is fixed up to better match the file format, and Song is updated to remove the data we're not interested in (i.e., in the qv-reference format, we're not interested in the quotes or the (qv) field).

A new unit test is added to cover the situation where there are multiple qv-referenced names in a field.