Test program: adding parsed song count and removing stopwatch.
The test program now outputs the total number of songs parsed using a LINQ query.

The stopwatch feature has been removed as the run-times for the parser are pretty short (~3 seconds to parse the whole file).  Based on the output from the test program, fewer than 200 songs fail to be parsed.